Final Assembly of CLU
The instructions implemented before were not working in the complete assembly they were working on separately. For the solution to this problem I designed a ring counter, the enables different paths of Microprocessors at per the requirement. I implemented all the instructions using the ring counter. Now all the instructions are working.